+++ lib/have_feature_pardiso_object.m
@@ -1,15 +1,19 @@
 function [TorF, vstr, rdate] = have_feature_pardiso_object()
 %HAVE_FEATURE_PARDISO_OBJECT  Detect availability/version info for PARDISO (object interface)
 %
-%   Used by HAVE_FEATURE.
+%   Feature detection function implementing 'pardiso_object' tag for
+%   HAVE_FEATURE to detect support for the object-oriented (v6.x and later)
+%   PARDISO interface.
+%
+%   See also HAVE_FEATURE, HAVE_FEATURE_PARDISO, HAVE_FEATURE_PARDISO_LEGACY.
 
-%   MP-Opt-Model
+%   MIPS
 %   Copyright (c) 2004-2020, Power Systems Engineering Research Center (PSERC)
 %   by Ray Zimmerman, PSERC Cornell
 %
-%   This file is part of MP-Opt-Model.
+%   This file is part of MIPS.
 %   Covered by the 3-clause BSD License (see LICENSE file for details).
-%   See https://github.com/MATPOWER/mp-opt-model for more info.
+%   See https://github.com/MATPOWER/mips for more info.
 
 TorF = exist('pardiso', 'file') == 2;
 vstr = '';